Refactor validators structure / update specs
