+++ app/validators/common/schemas/only_id_param_schema.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Common
+  module Schemas
+    OnlyIdParamSchema = Dry::Schema.Params do
+      required(:id).filled(:string)
+    end
+  end
+end
