--- app/validators/common/schemas/to_do_lists/create_to_do_list_schema.rb
@@ -1,12 +1,0 @@
-# frozen_string_literal: true
-
-module Common
-  module Schemas
-    module ToDoLists
-      CreateToDoListSchema = Dry::Schema.Params do
-        optional(:title).maybe(:string)
-        optional(:description).maybe(:string)
-      end
-    end
-  end
-end
