--- app/validators/common/schemas/to_do_lists/show_to_do_list_schema.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module Common
-  module Schemas
-    module ToDoLists
-      ShowToDoListSchema = Dry::Schema.Params do
-        required(:id).filled(:string)
-      end
-    end
-  end
-end
