+++ app/validators/to_do_lists/show_params_validator.rb
@@ -2,7 +2,7 @@
 
 module ToDoLists
   class ShowParamsValidator < Common::BaseValidator
-    params(Common::Schemas::ToDoLists::ShowToDoListSchema)
+    params(Common::Schemas::OnlyIdParamSchema)
 
     register_macro(:uuid_format) do
       key.failure(:invalid_format, field: :id, format_name: :UUID) unless UUID_FORMAT.match?(value)
